Order items now populated from purchase orders
Items dispatch will update dispatched status to true.

+++ database/migrations/2023_05_25_223034_create_purchase_order_details_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('reference')->nullable();
             $table->float('price');
             $table->float('quantity');
-            $table->float('total_price');
+            $table->double('total_price');
 
             $table->foreign('purchase_order_id')->references('id')->on('purchase_orders');
             $table->foreign('item_id')->references('id')->on('items');
